# apps/server/src/modules/manage/user/dto/edit-user.dto.ts
import {
  IsOptional,
  IsString,
  IsEmail,
  IsUUID,
  IsBoolean,
} from 'class-validator';
import { OmitType, PartialType } from '@nestjs/mapped-types';
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';

export class CreateUserDto {
  @ApiProperty({ description: '用户名' })
  @IsString()
  username: string;

  @ApiPropertyOptional({ description: '邮箱' })
  @IsOptional()
  @IsEmail()
  email?: string;

  @ApiPropertyOptional({ description: '手机号' })
  @IsOptional()
  @IsString()
  phone?: string;

  @ApiPropertyOptional({ description: '真实姓名' })
  @IsOptional()
  @IsString()
  realName?: string;

  @ApiPropertyOptional({ description: '头像 URL' })
  @IsOptional()
  @IsString()
  avatar?: string;

  @ApiProperty({ description: '密码' })
  @IsString()
  password: string;

  @ApiPropertyOptional({ description: '是否激活', default: true })
  @IsOptional()
  @IsBoolean()
  isActive?: boolean;
}

export class UpdateUserDto extends PartialType(
  OmitType(CreateUserDto, ['password']),
) {
  @ApiProperty({ description: '用户ID' })
  @IsString()
  @IsUUID()
  id: string;
}

export class UpdatePasswordDto {
  @ApiProperty({ description: '用户ID' })
  @IsString()
  @IsUUID()
  id: string;

  @ApiProperty({ description: '新密码' })
  @IsString()
  password: string;
}