# apps/server/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/manage/users/list": {
      "get": {
        "operationId": "UserController_list",
        "parameters": [
          {
            "name": "username",
            "required": false,
            "in": "query",
            "description": "用户名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "邮箱",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "description": "手机号",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "realName",
            "required": false,
            "in": "query",
            "description": "真实姓名",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "是否激活",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户列表"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取用户列表",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/manage/users/page": {
      "get": {
        "operationId": "UserController_page",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取用户分页数据"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "分页获取用户列表",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/manage/users/detail/{id}": {
      "get": {
        "operationId": "UserController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取用户详情"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取用户详情",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/manage/users/create": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "用户创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "创建用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/manage/users/update": {
      "post": {
        "operationId": "UserController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "用户更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/manage/users/delete": {
      "post": {
        "operationId": "UserController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "用户删除成功"
          },
          "400": {
            "description": "删除用户失败"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "删除用户",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/manage/users/updatePassword": {
      "post": {
        "operationId": "UserController_updatePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "密码更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "用户不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新用户密码",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/manage/users/grant": {
      "post": {
        "operationId": "UserController_grant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "授权成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "用户或角色不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "授予用户角色",
        "tags": [
          "用户管理"
        ]
      }
    },
    "/api/v1/manage/users/revoke": {
      "post": {
        "operationId": "UserController_revoke",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "撤销授权成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "用户或角色不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "撤销用户角色",
        "tags": [
          "用户管理"
        ]
      }
    },
...
    "/api/v1/manage/permission/list": {
      "get": {
        "operationId": "PermissionController_findAll",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "权限名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "权限编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "required": false,
            "in": "query",
            "description": "权限描述",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "是否激活",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取权限列表"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取权限列表",
        "tags": [
          "权限管理"
        ]
      }
    },
    "/api/v1/manage/permission/page": {
      "get": {
        "operationId": "PermissionController_findPage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取权限分页数据"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "分页获取权限列表",
        "tags": [
          "权限管理"
        ]
      }
    },
    "/api/v1/manage/permission/detail/{id}": {
      "get": {
        "operationId": "PermissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取权限详情"
          },
          "404": {
            "description": "权限不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取权限详情",
        "tags": [
          "权限管理"
        ]
      }
    },
    "/api/v1/manage/permission/create": {
      "post": {
        "operationId": "PermissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "权限创建成功"
          },
          "400": {
            "description": "请求参数错误"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "创建权限",
        "tags": [
          "权限管理"
        ]
      }
    },
    "/api/v1/manage/permission/update": {
      "post": {
        "operationId": "PermissionController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "权限更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "404": {
            "description": "权限不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "更新权限",
        "tags": [
          "权限管理"
        ]
      }
    },
    "/api/v1/manage/permission/delete": {
      "post": {
        "operationId": "PermissionController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "权限删除成功"
          },
          "400": {
            "description": "删除权限失败"
          },
          "404": {
            "description": "权限不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "删除权限",
        "tags": [
          "权限管理"
        ]
      }
    },
    "/api/v1/manage/permission/rolePermissions/{roleId}": {
      "get": {
        "operationId": "PermissionController_getRolePermissions",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取角色权限"
          },
          "404": {
            "description": "角色不存在"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "获取角色权限",
        "tags": [
          "权限管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-Commerce API",
    "description": "API documentation for the e-commerce management system",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "realName": {
            "type": "string",
            "description": "真实姓名"
          },
          "avatar": {
            "type": "string",
            "description": "头像 URL"
          },
          "password": {
            "type": "string",
            "description": "密码"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否激活",
            "default": true
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "用户ID"
          },
          "password": {
            "type": "string",
            "description": "新密码"
          }
        },
        "required": [
          "id",
          "password"
        ]
      },
      "GrantUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "用户ID"
          },
          "roleIds": {
            "description": "角色ID列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "roleIds"
        ]
      },
      "RevokeUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "用户ID"
          },
          "roleIds": {
            "description": "角色ID列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "roleIds"
        ]
      },
...
      "CreatePermissionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "权限名称",
            "maxLength": 50
          },
          "code": {
            "type": "string",
            "description": "权限编码",
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "description": "权限描述",
            "maxLength": 255
          },
          "isActive": {
            "type": "boolean",
            "description": "是否激活",
            "default": true
          }
        },
        "required": [
          "name",
          "code"
        ]
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "权限ID"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}